pyproject.toml
[project]
name = "sparsemap"
version = "0.1.0"
description = "High-signal knowledge graph extractor. Mapping the critical path through the noise."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.3",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.128.0",
    "google-genai>=1.60.0",
    "httpx>=0.28.1",
    "openai>=2.16.0",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
]

[project.scripts]
sparsemap = "sparsemap.main:main"

[dependency-groups]
dev = [
    "ruff>=0.14.14",
]

[build-system]
requires = ["uv_build>=0.9.28,<0.10.0"]
build-backend = "uv_build"
